Dev jcy

+++ pizza-data/src/main/java/edu.ecnu.scsse.pizza.data/repository/PizzaShopJpaRepository.java
@@ -3,6 +3,6 @@ package edu.ecnu.scsse.pizza.data.repository;
 import edu.ecnu.scsse.pizza.data.domain.PizzaShopEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PizzaShopJpaRepository extends JpaRepository<PizzaShopEntity,Long> {
+public interface PizzaShopJpaRepository extends JpaRepository<PizzaShopEntity,Integer> {
 
 }
